requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
altair==4.2.2
anyascii==0.3.2
async-timeout==5.0.1
attrs==24.2.0
audioread==3.0.1
babel==2.16.0
bangla==0.0.2
blinker==1.9.0
blis==0.7.11
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
confection==0.1.5
contourpy==1.2.1
coqpit==0.0.17
cycler==0.12.1
cymem==2.0.10
Cython==3.0.11
dateparser==1.1.8
decorator==4.4.2
docopt==0.6.2
einops==0.8.0
encodec==0.1.1
entrypoints==0.4
ffmpeg==1.4
filelock==3.16.1
Flask==3.1.0
fonttools==4.55.2
frozenlist==1.5.0
fsspec==2024.10.0
g2pkk==0.1.2
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.68.1
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
hangul-romanize==0.1.0
huggingface-hub==0.26.5
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.5.0
inflect==7.4.0
itsdangerous==2.2.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.4
joblib==1.4.2
jsonlines==1.2.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
librosa==0.10.0
llvmlite==0.43.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.8.4
mdurl==0.1.2
more-itertools==10.5.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
murmurhash==1.0.11
narwhals==1.15.2
networkx==2.8.8
nltk==3.9.1
num2words==0.5.13
numba==0.60.0
numpy==1.22.0
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai-whisper==20240930
packaging==24.2
pandas==1.5.3
pathlib_abc==0.1.1
pathy==0.11.0
pillow==11.0.0
platformdirs==4.3.6
pooch==1.8.2
preshed==3.0.9
proglog==0.1.10
propcache==0.2.1
protobuf==3.20.3
psutil==6.1.0
pyarrow==18.1.0
pycparser==2.22
pydantic==1.10.19
pydeck==0.9.1
Pygments==2.18.0
Pympler==1.1
pynndescent==0.5.13
pyparsing==3.2.0
pypinyin==0.53.0
pysbd==0.3.4
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.11.4
semver==3.0.2
sentencepiece==0.2.0
six==1.17.0
smart-open==6.4.0
smmap==5.0.1
soundfile==0.12.1
soxr==0.5.0.post1
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
streamlit==1.18.0
SudachiDict-core==20241021
SudachiPy==0.6.9
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
thinc==8.1.12
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
toml==0.10.2
toolz==1.0.0
torch==2.5.1
torchaudio==2.5.1
tornado==6.4.2
tqdm==4.67.1
trainer==0.0.36
transformers==4.47.0
triton==3.1.0
TTS==0.22.0
typeguard==4.4.1
typer==0.7.0
typing_extensions==4.12.2
tzlocal==5.2
umap-learn==0.5.7
Unidecode==1.3.8
urllib3==2.2.3
validators==0.34.0
wasabi==1.1.3
watchdog==6.0.0
Werkzeug==3.1.3
yarl==1.18.3
zipp==3.21.0
